faithful: BYTE-IDENTICAL to LCP_ORG.PRG (MD5 02900cfd)
Phase 2/3 complete -- the -DFAITHFUL build now reproduces the 1985
binary bit-for-bit (87832 bytes, text/data/bss/reloc all identical).

Data layout (globals.c and friends reordered to ROM declaration
order):
- mi_evi/mi_evcn(=9)/mi_nlp0(=100) are initialized data at 0x120fa;
  g_msmk block follows g_msmsa; env_val block follows psg_ntAc.
- Kept-engine sequencer state (mi_*/psg_* tables) gated out of
  FAITHFUL: none of it exists in the ROM's data segment.
- Minigame window at 0x12484: g_agwgm before g_agorw; g_aggpr and
  poker/wordpuzzle strings/tables are kept-only; g_pcbet/g_ppbet
  precede g_pcmon at 0x124a0 (pk_main's clears reference them).
- g_ptdsi is 12 shorts at 0x1358a (ping-pong + 0 terminator),
  g_obdea is [3]; g_cotbl is 19 shorts of {id, 0} pairs.
- g_dsb statically points at dsb_stor (reloc'd data 0x1214e).
- cs_mvIn sets lcp_face, not g_lcyof (ROM 0x8106 truth).

BSS layout:
- Kept-only commons gated (#ifndef FAITHFUL); ROM sizes restored
  under FAITHFUL: SPRITE_HW_SLOTS_ALLOC=8, pst_arr[4],
  dsb_stor[16256], pk_ch/pk_ph[26] (war hands), assets
  body_buf[20000]/pex_buf[12000].
- The ROM stores the loaded PEx frame pointer over pex_name's first
  4 bytes; FAITHFUL aliases pex_ptr onto _pex_nam via Alcyon's
  8-char symbol truncation (extern short *pex_namP).
- The 1985 linker's .comm allocation order matches no surviving
  linker (lo68/link68 hash-group, ALN.PRG sorts alphabetically,
  and it is neither first- nor last-mention order).  New
  tools/bss_remap.py pairs the byte-identical relocation streams
  site-by-site, verifies the port->ROM BSS translation is a
  consistent 1:1 mapping (222 addresses, 873 sites), and applies
  it as the final FAITHFUL link step.

prg_diff.py gains --mask (zero relocated longs) used throughout the
campaign.  Kept build unaffected; 8000-VBL Hatari smoke clean.

+++ source/include/tick_tables.h
@@ -11,7 +11,7 @@ extern short g_obpha[];
 extern short g_obfia[];
 extern short g_obdea[];
 extern short g_ptdsi[];
-extern long g_cotbl[];
+extern short g_cotbl[];
 extern short g_ptlss;
 extern BOOL16 g_alsts;
 extern short g_phrc;
